reorganized file and data structure

+++ server/entity.h
@@ -1,12 +1,12 @@
 #ifndef ENTITY_H
 #define ENTITY_H
 typedef struct entity{
-	double center[3];
-	double testcenter[3];
-	double vel[3];
-	double testvel[3];
-	double mass;
-	double radius;
+	int center[3];
+	int testcenter[3];
+	int vel[3];
+	int testvel[3];
+	int mass;
+	int radius;
 	char keys;
 }entity;
 extern entity* entityList;
